organize config folder moar
