+++ src/api/queries/album/utils/album.ts
@@ -9,7 +9,7 @@ import { JellifyLibrary } from '../../../../types/JellifyLibrary'
 import { Api } from '@jellyfin/sdk'
 import { fetchItem } from '../../item'
 import { JellifyUser } from '../../../../types/JellifyUser'
-import { ApiLimits } from '../../../../configs/query.config'
+import { ApiLimits } from '../../../../configs/querying/index.config'
 import buildYearsParam from '../../../../utils/mapping/build-years-param'
 import { getItemsApi } from '@jellyfin/sdk/lib/utils/api/items-api'
 import { setQueryUserDataForItems } from '../../user-data'
