+++ src/api/queries/genre/utils/index.ts
@@ -3,7 +3,7 @@ import { Api } from '@jellyfin/sdk'
 import { JellifyLibrary } from '../../../../types/JellifyLibrary'
 import { JellifyUser } from '../../../../types/JellifyUser'
 import { isUndefined } from 'lodash'
-import { ApiLimits } from '../../../../configs/query.config'
+import { ApiLimits } from '../../../../configs/querying/index.config'
 import { ItemSortBy } from '@jellyfin/sdk/lib/generated-client/models/item-sort-by'
 import { SortOrder } from '@jellyfin/sdk/lib/generated-client/models/sort-order'
 import { ItemFields } from '@jellyfin/sdk/lib/generated-client/models'
