+++ src/api/queries/instant-mix/utils/index.ts
@@ -1,7 +1,7 @@
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client/models'
 import { getInstantMixApi } from '@jellyfin/sdk/lib/utils/api'
 import { isUndefined } from 'lodash'
-import QueryConfig from '../../../../configs/query.config'
+import QueryConfig from '../../../../configs/querying/index.config'
 import { Api } from '@jellyfin/sdk'
 import { JellifyUser } from '../../../../types/JellifyUser'
 /**
