+++ src/api/queries/playlist/utils/index.ts
@@ -9,7 +9,7 @@ import { getItemsApi } from '@jellyfin/sdk/lib/utils/api'
 import { JellifyUser } from '../../../../types/JellifyUser'
 import { Api } from '@jellyfin/sdk'
 import { isUndefined } from 'lodash'
-import QueryConfig, { ApiLimits } from '../../../../configs/query.config'
+import QueryConfig, { ApiLimits } from '../../../../configs/querying/index.config'
 import { setQueryUserDataForItems } from '../../user-data'
 import { ensurePlaylistLibraryQueryData } from '../../libraries'
 import { captureError, LoggingContext } from '../../../../utils/logging'
