+++ src/api/queries/recents/utils/index.ts
@@ -7,14 +7,14 @@ import {
 	SortOrder,
 } from '@jellyfin/sdk/lib/generated-client/models'
 import { getItemsApi } from '@jellyfin/sdk/lib/utils/api/items-api'
-import { ApiLimits } from '../../../../configs/query.config'
+import { ApiLimits } from '../../../../configs/querying/index.config'
 import { getUserLibraryApi } from '@jellyfin/sdk/lib/utils/api'
 import { Api } from '@jellyfin/sdk'
 import { isUndefined } from 'lodash'
 import { JellifyLibrary } from '../../../../types/JellifyLibrary'
 import { JellifyUser } from '../../../../types/JellifyUser'
 import { queryClient } from '../../../../constants/query-client'
-import { RECENTLY_PLAYED_ALBUM_THRESHOLD } from '../../../../configs/home.config'
+import { RECENTLY_PLAYED_ALBUM_THRESHOLD } from '../../../../configs/categorizing/home.config'
 import { PlayItAgainQuery } from '..'
 import { ArtistQueryKey } from '../../artist/keys'
 import { setQueryUserDataForItem } from '../../user-data'
