+++ src/api/queries/search/utils/index.ts
@@ -1,7 +1,7 @@
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client/models'
 import { getItemsApi } from '@jellyfin/sdk/lib/utils/api'
 import { isEmpty, isUndefined, trim } from 'lodash'
-import QueryConfig from '../../../../configs/query.config'
+import QueryConfig from '../../../../configs/querying/index.config'
 import { getApi, getUser } from '../../../../stores/auth/utils'
 import { setQueryUserDataForItems } from '../../user-data'
 /**
