+++ src/configs/categorizing/home.config.ts
