+++ src/constants/query-client.ts
@@ -1,5 +1,5 @@
 import { QueryClient } from '@tanstack/react-query'
-import { MAX_RETRY_ATTEMPTS } from '../configs/query.config'
+import { MAX_RETRY_ATTEMPTS } from '../configs/querying/index.config'
 
 export const ONE_MINUTE = 1000 * 60
 export const ONE_HOUR = ONE_MINUTE * 60
